Drop Operation.move
Amount of additional logic it requires across the library is crazy.

It is definitely possible to handle, but at a cost of greatly increased complexity.

In light that move can always be achieved by combination of remove + insert, I’m dropping move support for now.

Maybe in the future major upgrade…